pdf2cbz/pyproject.toml
[project]
name = "pdf2cbz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
]
